add token

+++ Routers/ManageExperiment/stuExperiment.py
@@ -25,7 +25,7 @@ def createFilePath(sID,exID,filename):
     isExists=os.path.exists(path)
     if not isExists:
         os.makedirs(path) 
-    newFileName = filename + '_' + sID + str(uuid.uuid1()) + ext
+    newFileName = sID + '_' + filename  + str(uuid.uuid1()) + ext
     path = os.path.join(path,newFileName)
     return path
     
@@ -39,14 +39,16 @@ def uoloadReport():
     path = createFilePath(s_id,ex_id,report.filename)
     se = StudentExperiment.query.filter(StudentExperiment.experiment_id == ex_id,StudentExperiment.s_id == s_id).first()
     if se:
-        os.remove(se.file_url)
-        se.file_url = path
-    else :
-        se = StudentExperiment(experiment_id = ex_id, s_id = s_id, file_url = path)
-        dbManage.db.session.add(se)
-    report.save(path)
-    dbManage.db.session.commit()
-    return "success"
+        if se.file_url:
+            os.remove(se.file_url)
+            se.file_url = path
+        else :
+            se.file_url = path
+        report.save(path)
+        dbManage.db.session.commit()
+        return "success"
+    else:
+        return "NotExist"
 
 @studentExperimentRoute.route('/Ex/getUpload/',methods=['POST'])  
 def getReport():
@@ -83,8 +85,13 @@ def showEx():
     Exs = Experiment.query.filter(Experiment.course_id == course_id).all()
     content = []
     for ex in Exs:
+        now_time = datetime.datetime.now()
+        if now_time > ex.end_time:
+            status = 3
+        else:
+            status = ex.status
         se = StudentExperiment.query.filter(StudentExperiment.s_id == s_id, StudentExperiment.experiment_id == ex.experiment_id).first()
-        temp = {"ex_id": ex.experiment_id, "experiment_title":ex.experiment_title, "experiment_brief":ex.experiment_brief, "end_time":ex.end_time, "weight":ex.weight, "score": se.score}
+        temp = {"ex_id": ex.experiment_id, "experiment_title":ex.experiment_title, "experiment_brief":ex.experiment_brief, "end_time":str(ex.end_time), "weight":ex.weight, "score": se.score,"status": status}
         content.append(temp)
     return jsonify(content)
